+++ app/src/main/java/com/chydee/notekeeper/dao/NoteDao.java
@@ -1,4 +1,4 @@
-package com.chydee.notekeeper;
+package com.chydee.notekeeper.dao;
 
 import androidx.lifecycle.LiveData;
 import androidx.room.Dao;
